refactor(siiges): refactor siiges for build
