+++ apps/siiges-app/mocks/browser.js
@@ -1,4 +1,6 @@
 import { setupWorker } from 'msw';
-import { handlers } from './handlers';
+import handlers from './handler';
 
-export const worker = setupWorker(...handlers);
+const worker = setupWorker(...handlers);
+
+export default worker;
