+++ apps/siiges-app/mocks/handler.js
@@ -5,18 +5,14 @@ const coins = {
   DOT: { name: 'Polkadot', website: 'https://polkadot.network' },
 };
 
-export const handlers = [
-  rest.get('/coins', (req, res, ctx) => {
-    return res(
-      ctx.status(200),
-      ctx.delay(500),
-      ctx.json(
-        Object.entries(coins).map(([symbol, data]) => {
-          return { symbol, ...data };
-        })
-      )
-    );
-  }),
+const handlers = [
+  rest.get('/coins', (req, res, ctx) => res(
+    ctx.status(200),
+    ctx.delay(500),
+    ctx.json(
+      Object.entries(coins).map(([symbol, data]) => ({ symbol, ...data })),
+    ),
+  )),
 
   rest.get('https://coins.com/:symbol', (req, res, ctx) => {
     const symbol = req.params.symbol.toUpperCase();
@@ -29,3 +25,5 @@ export const handlers = [
     return res(ctx.json({ symbol, ...data }));
   }),
 ];
+
+export default handlers;
