+++ apps/siiges-app/package.json
@@ -14,6 +14,7 @@
     "@siiges-ui/authentication": "^1.0.0",
     "@siiges-ui/inspecciones": "^1.0.0",
     "@siiges-ui/instituciones": "^1.0.0",
+    "@siiges-ui/serviciosescolares": "^1.0.0",
     "@siiges-ui/shared": "1.0.0",
     "@siiges-ui/solicitudes": "^1.0.0",
     "@siiges-ui/users": "^1.0.0",
