+++ apps/siiges-app/pages/institucion/[institucionId]/consultarPlantel/[plantelId].jsx
@@ -1,7 +1,7 @@
 import { ConsultPlantelesForm } from '@siiges-ui/instituciones';
 import { Layout } from '@siiges-ui/shared';
 import React from 'react';
-import getPlantel from '../../utils/getPlantel';
+import getPlantel from '@siiges-ui/instituciones/src/utils/getPlantel';
 
 export default function consultarPlantel() {
   const { plantel, loading } = getPlantel();
