+++ apps/siiges-app/pages/institucion/[institucionId]/editarPlantel/[plantelId].jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Layout } from '@siiges-ui/shared';
 import { NewPlantelForm } from '@siiges-ui/instituciones';
-import getPlantel from '../../utils/getPlantel';
+import getPlantel from '@siiges-ui/instituciones/src/utils/getPlantel';
 
 export default function EditarPlantel() {
   const { plantel, loading } = getPlantel();
