+++ apps/siiges-app/pages/serviciosEscolares/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Layout, DataTable } from '@siiges-ui/shared';
 import { Grid } from '@mui/material';
-import columns from './Tables/institucionesAutorizadas';
-import getInstituciones from '../institucion/utils/getInstituciones';
+import getInstituciones from '@siiges-ui/instituciones/src/utils/getInstituciones';
+import columns from '@siiges-ui/serviciosescolares';
 
 export default function serviciosEscolares() {
   const { instituciones, loading } = getInstituciones();
