+++ apps/siiges-app/pages/usuarios/editarUsuario/[id].jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Layout } from '@siiges-ui/shared';
-import { EditUserForm } from '@siiges-ui/users';
+import { EditUserForm, getUser } from '@siiges-ui/users';
 import Grid from '@mui/material/Grid';
-import getUser from '../utils/getUser';
 
 export default function EditUser() {
   const { user, loading } = getUser();
