+++ packages/serviciosEscolares/README.md
@@ -1,0 +1,11 @@
+# `@siiges-ui/serviciosEscolares`
+
+> TODO: description
+
+## Usage
+
+```
+const serviciosescolares = require('@siiges-ui/serviciosEscolares');
+
+// TODO: DEMONSTRATE API
+```
