+++ packages/serviciosEscolares/__tests__/serviciosEscolares.test.js
@@ -1,0 +1,3 @@
+describe('@siiges-ui/serviciosescolares', () => {
+  it.todo('needs tests');
+});
