+++ packages/serviciosEscolares/src/Tables/institucionesAutorizadas.jsx
