+++ packages/serviciosEscolares/src/index.js
@@ -1,0 +1,7 @@
+import columns from './Tables/institucionesAutorizadas';
+import ButtonsInstitucionesAutorizadas from './Components/ButtonsInstitucionesAutorizadas';
+
+export {
+  columns,
+  ButtonsInstitucionesAutorizadas,
+};
