+++ packages/users/src/index.js
@@ -3,6 +3,10 @@ import UserConsult from './components/info/UserConsult';
 import UserForm from './components/forms/ConsultUserForm';
 import UserInfo from './components/info/UserInfo';
 import EditUserForm from './components/forms/EditUserForm';
+import deleteUser from './utils/deleteUser';
+import getUsers from './utils/getUsers';
+import getUser from './utils/getUser';
+import columns from './Mocks/usuarios';
 
 export {
   NewUserForm,
@@ -10,4 +14,8 @@ export {
   UserConsult,
   UserForm,
   UserInfo,
+  deleteUser,
+  getUsers,
+  getUser,
+  columns,
 };
